[IMPROVE] DI and Constant
