+++ Telemetry Device/Models/Constant/ConstantBits.cs
@@ -13,5 +13,7 @@ namespace SendRecieveUDP.Model.Constant
         public const int SINGLE_BIT_VALUE = 1;
         public const int NO_OFFSET = 0;
         public const int WORD = 16;
+        public const int WORD_SIZE_BYTES = 2;
+
     }
 }
