+++ Telemetry Device/Models/Interface/Networking/IUdpChecksumCalculator.cs
@@ -1,0 +1,7 @@
+﻿namespace Telemetry_Device.Models.Interface.Networking
+{
+    public interface IUdpChecksumCalculator
+    {
+        ushort ComputeUdpChecksum(byte[] frame);
+    }
+}
