+++ Telemetry Device/Services/TplDataflowBlocks/BuilderBlock.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks.Dataflow;
 using Telemetry_Device.Models.Constant;
+using Telemetry_Device.Models.Interface.Networking;
 using Telemetry_Device.Models.Interface.TplBlocks;
 using Telemetry_Device.Models.Packets;
 
@@ -11,6 +12,12 @@ namespace Telemetry_Device.Services.TplDataflowBlocks
 {
     public class BuilderBlock : IBuilderBlock
     {
+        private readonly IUdpChecksumCalculator _udpChecksumCalculator;
+        public BuilderBlock(IUdpChecksumCalculator checksumCalculator)
+        {
+            _udpChecksumCalculator = checksumCalculator;
+        }
+
         public BufferBlock<PacketData> CreateBuilderBlock(string filePath)
         {
             BufferBlock<PacketData> buffer = new BufferBlock<PacketData>();
@@ -25,7 +32,7 @@ namespace Telemetry_Device.Services.TplDataflowBlocks
                 while (binaryReader.BaseStream.Position < binaryReader.BaseStream.Length)
                 {
                     PacketData packet = ReadSinglePacket(binaryReader);
-                    if (ComputeUdpChecksum(packet.PayloadPacket) == packet.ChecksumPacket)
+                    if (_udpChecksumCalculator.ComputeUdpChecksum(packet.PayloadPacket) == packet.ChecksumPacket)
                     {
                         buffer.Post(packet);
                     }
@@ -56,43 +63,6 @@ namespace Telemetry_Device.Services.TplDataflowBlocks
                 ChecksumPacket = checksum
             };
         }
-
-
-
-
-
-
-
-        public ushort ComputeUdpChecksum(byte[] frame)
-        {
-            uint checksum = 0;
-            for (int offset = ConstantPackets.START_IP_HEADER; offset <= ConstantPackets.END_IP_HEADER; offset += 2)
-                AddWord(ref checksum, (ushort)((frame[ConstantPackets.FRAME_IP_HEADER_OFFSET + offset] << ConstantBits.BITS_IN_BYTE)
-                                             | frame[ConstantPackets.FRAME_IP_HEADER_OFFSET + offset + 1]));
-
-            AddWord(ref checksum, ConstantNetwork.UDP_PROTOCOL_NUMBER);
-            AddWord(ref checksum, ConstantPackets.FIXED_UDP_LENGTH);
-
-            for (int udpWordOffset = 0; udpWordOffset < ConstantPackets.FIXED_UDP_LENGTH; udpWordOffset += 2)
-            {
-                if (udpWordOffset != ConstantPackets.CHECKSUM_OFFSET)
-                {
-                    int frameByteIndex = ConstantPackets.FRAME_UDP_HEADER_OFFSET + udpWordOffset;
-                    byte highByte = frame[frameByteIndex];
-                    byte lowByte = (udpWordOffset + 1 < ConstantPackets.FIXED_UDP_LENGTH) ? frame[frameByteIndex + 1] : (byte)0;
-
-                    AddWord(ref checksum, (ushort)((highByte << ConstantBits.BITS_IN_BYTE) | lowByte));
-                }
-            }
-            return (ushort)~checksum;
-        }
-
-        public void AddWord(ref uint sum, ushort word)
-        {
-            sum += word;
-            sum = (sum & 0xFFFF) + (sum >> ConstantBits.WORD);
-        }
-
     }
 
 }
